+++ packages/orq-rc/src/lib/config.ts
@@ -68,7 +68,7 @@ export function serverURLFromOptions(options: SDKOptions): URL | null {
 export const SDK_METADATA = {
   language: "typescript",
   openapiDocVersion: "2.0",
-  sdkVersion: "4.1.0-rc.44",
-  genVersion: "2.781.2",
-  userAgent: "speakeasy-sdk/typescript 4.1.0-rc.44 2.781.2 2.0 @orq-ai/node",
+  sdkVersion: "4.1.0-rc.45",
+  genVersion: "2.785.0",
+  userAgent: "speakeasy-sdk/typescript 4.1.0-rc.45 2.785.0 2.0 @orq-ai/node",
 } as const;