+++ packages/orq-rc/src/models/operations/createbudget.ts
@@ -489,7 +489,7 @@ export const CreateBudgetResponseBody$inboundSchema: z.ZodType<
   created: z.string().datetime({ offset: true }).transform(v => new Date(v))
     .optional(),
   updated: z.string().datetime({ offset: true }).default(
-    "2025-12-16T05:40:52.364Z",
+    "2025-12-16T10:45:03.543Z",
   ).transform(v => new Date(v)),
 }).transform((v) => {
   return remap$(v, {
@@ -529,7 +529,7 @@ export const CreateBudgetResponseBody$outboundSchema: z.ZodType<
   isActive: z.boolean(),
   consumption: z.lazy(() => Consumption$outboundSchema).optional(),
   created: z.date().transform(v => v.toISOString()).optional(),
-  updated: z.date().default(() => new Date("2025-12-16T05:40:52.364Z"))
+  updated: z.date().default(() => new Date("2025-12-16T10:45:03.543Z"))
     .transform(v => v.toISOString()),
 }).transform((v) => {
   return remap$(v, {