+++ packages/orq-rc/src/models/operations/generateconversationname.ts
@@ -248,7 +248,7 @@ export const GenerateConversationNameResponseBody$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  _id: z.string().default("conv_01kcjttd2w2qgamm07nz07f9ss"),
+  _id: z.string().default("conv_01kckc7b37pm4xew3swnf4r6x0"),
   entityId: z.string(),
   kind: GenerateConversationNameKind$inboundSchema,
   displayName: z.string(),
@@ -282,7 +282,7 @@ export const GenerateConversationNameResponseBody$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   GenerateConversationNameResponseBody
 > = z.object({
-  id: z.string().default("conv_01kcjttd2w2qgamm07nz07f9ss"),
+  id: z.string().default("conv_01kckc7b37pm4xew3swnf4r6x0"),
   entityId: z.string(),
   kind: GenerateConversationNameKind$outboundSchema,
   displayName: z.string(),