+++ packages/orq-rc/src/models/operations/streamrunagent.ts
@@ -4091,7 +4091,7 @@ export const AgentToolInputRunTools$inboundSchema: z.ZodType<
   z.ZodTypeDef,
   unknown
 > = z.object({
-  id: z.string().default("01KCJTTD0ZHYMQRWMCHQS9XYQR"),
+  id: z.string().default("01KCKC7B1B6B35T85M8EHJ3T1R"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$inboundSchema),
@@ -4110,7 +4110,7 @@ export const AgentToolInputRunTools$outboundSchema: z.ZodType<
   z.ZodTypeDef,
   AgentToolInputRunTools
 > = z.object({
-  id: z.string().default("01KCJTTD0ZHYMQRWMCHQS9XYQR"),
+  id: z.string().default("01KCKC7B1B6B35T85M8EHJ3T1R"),
   name: z.string(),
   description: z.string().optional(),
   schema: z.lazy(() => AgentToolInputRunSchema$outboundSchema),