+++ packages/orq-rc/RELEASES.md
@@ -5599,3 +5599,13 @@ Based on:
 - [typescript v4.1.0-rc.44] packages/orq-rc
 ### Releases
 - [NPM v4.1.0-rc.44] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.44 - packages/orq-rc
+
+## 2025-12-16 10:47:17
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.677.0 (2.785.0) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v4.1.0-rc.45] packages/orq-rc
+### Releases
+- [NPM v4.1.0-rc.45] https://www.npmjs.com/package/@orq-ai/node/v/4.1.0-rc.45 - packages/orq-rc